Move delete button into the header

+++ app/javascript/components/bootcamp/CustomFunctionEditor/DeleteFunctionButton.tsx
@@ -1,32 +1,36 @@
 import React, { useCallback, useContext } from 'react'
-import { SolveExercisePageContext } from '../SolveExercisePage/SolveExercisePageContextWrapper'
-import { redirectTo } from '@/utils/redirect-to'
 import toast from 'react-hot-toast'
-export function DeleteFunctionButton() {
+import { redirectTo } from '@/utils/redirect-to'
+import { SolveExercisePageContext } from '../SolveExercisePage/SolveExercisePageContextWrapper'
+
+export function DeleteFunctionButton({ predefined }: { predefined: boolean }) {
   const { links } = useContext(SolveExercisePageContext)
 
   const handleDeleteCustomFunction = useCallback(() => {
     const url = links.deleteCustomFn as string
     const customFnDashboardUrl = links.customFnsDashboard as string
 
-    deleteCustomFunction(url)
-      .then(() => {
-        toast.success(
-          'Deleted custom function successfully. Redirecting you to the dashboard..'
-        )
-        setTimeout(() => redirectTo(customFnDashboardUrl), 1000)
-      })
-      .catch((e) => {
-        toast.error("Couldn't delete custom function. Please try it again.")
-      })
+    if (window.confirm('Are you sure you want to delete this function?')) {
+      deleteCustomFunction(url)
+        .then(() => {
+          toast.success(
+            'Deleted custom function successfully. Redirecting you to the dashboard..'
+          )
+          setTimeout(() => redirectTo(customFnDashboardUrl), 1000)
+        })
+        .catch((e) => {
+          toast.error("Couldn't delete custom function. Please try it again.")
+        })
+    }
   }, [links])
 
+  if (predefined) {
+    return null
+  }
+
   return (
-    <button
-      onClick={handleDeleteCustomFunction}
-      className="btn-alert btn-s justify-self-end mt-24"
-    >
-      Delete library function
+    <button onClick={handleDeleteCustomFunction} className="btn-alert btn-xxs">
+      Delete
     </button>
   )
 }
